first form edit issue fix

+++ app/Http/Controllers/JobApplicationController.php
@@ -224,21 +224,22 @@ class JobApplicationController extends Controller
         $jobApplication = JobApplication::find($id);
         $jobApplication = $jobApplication->fill($data);
         $jobApplication->save();
-
-        if($request->has('siblings_name')) {
+        if($request->has('siblings_name') && $request->siblings == 'yes') {
             $jobApplication->siblingsList()->delete();
             foreach ($request->siblings_name as $key => $sibling) {
-                if (isset($request->file('siblings_image')[$key])) {
-                    $siblingsImagepath = $request->file('siblings_image')[$key]->store('images', 'public');
-                    $siblingImage = $siblingsImagepath;
+                if(!is_null($sibling)) {
+                    if (isset($request->file('siblings_image')[$key])) {
+                        $siblingsImagepath = $request->file('siblings_image')[$key]->store('images', 'public');
+                        $siblingImage = $siblingsImagepath;
+                    }
+                    $jobApplication->siblingsList()->create([
+                        'name' => $request->siblings_name[$key],
+                        'dob' => $request->siblings_date_of_birth[$key],
+                        'email' => $request->siblings_email[$key],
+                        'phone' => $request->siblings_phone[$key],
+                        'photo' => $siblingImage ?? $request->siblings_image_old[$key],
+                    ]);
                 }
-                $jobApplication->siblingsList()->create([
-                    'name' => $request->siblings_name[$key],
-                    'dob' => $request->siblings_date_of_birth[$key],
-                    'email' => $request->siblings_email[$key],
-                    'phone' => $request->siblings_phone[$key],
-                    'photo' => $siblingImage ?? $request->siblings_image_old[$key],
-                ]);
             }
         }
